backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
psycopg[binary]>=3.1
python-dotenv
pypdf
python-docx
pgvector
